test fixes, fcml_bool expanded to int

+++ fcml_types.h
@@ -11,7 +11,7 @@
 #include <stddef.h>
 #include <stdint.h>
 
-typedef uint8_t fcml_bool;
+typedef int fcml_bool;
 typedef int8_t fcml_int8_t;
 typedef uint8_t fcml_uint8_t;
 typedef int16_t fcml_int16_t;
